+++ arviz/__init__.py
@@ -25,8 +25,8 @@ if not logging.root.handlers:
 
 from .rcparams import rcParams, rc_context
 from .data import *
-from .plots import *
 from .stats import *
+from .plots import *
 from .utils import Numba, interactive_backend
 from .plots import backends
 from .wrappers import *